improve: AstMetaData 解耦合。

+++ src/main/java/cn/addenda/ro/grammar/ast/expression/TimeInterval.java
@@ -3,6 +3,8 @@ package cn.addenda.ro.grammar.ast.expression;
 import cn.addenda.ro.grammar.ast.CurdVisitor;
 import cn.addenda.ro.grammar.lexical.token.Token;
 
+import java.math.BigInteger;
+
 /**
  * @author addenda
  * @datetime 2021/7/26 23:16
@@ -11,9 +13,9 @@ public class TimeInterval extends Curd {
 
     private Token timeType;
 
-    private Long interval;
+    private BigInteger interval;
 
-    public TimeInterval(Token timeType, Long interval) {
+    public TimeInterval(Token timeType, BigInteger interval) {
         this.timeType = timeType;
         this.interval = interval;
     }
@@ -27,7 +29,7 @@ public class TimeInterval extends Curd {
         return timeType;
     }
 
-    public Long getInterval() {
+    public BigInteger getInterval() {
         return interval;
     }
